requirements.txt
absl-py==0.12.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cloudpickle==1.3.0
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.16
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
enum34==1.1.10
future==0.18.2
google-auth==1.30.1
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
imageio==2.8.0
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
littleutils==0.2.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
networkx==2.4
notebook==6.0.3
numpy==1.19.5
oauthlib==3.1.0
ogb==1.3.1
opencv-python==4.5.2.52
outdated==0.2.1
pandas==1.1.5
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pygobject==3.26.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-distutils-extra==2.39
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.8.3
testpath==0.4.4
torch==1.8.1
torch-scatter==2.0.6
torchvision==0.9.1
tornado==6.0.4
tqdm==4.61.0
traitlets==4.3.3
typing==3.7.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wilds==1.1.0
zipp==3.1.0
